move `custom-event-target.ts`
