+++ src/custom-event-target.ts
