+++ src/player_api/index.ts
@@ -1,3 +1,3 @@
 export * from './yt-api';
 export * from './manager';
-export type { EventMapOf } from './custom-event-target';
+export type { EventMapOf } from '../custom-event-target';
